Solution to Input and Print a Double

+++ Input and Print a Double@@6788d5012db25efec946f942/code.c
@@ -1,8 +1,8 @@
 #include <stdio.h>
 
 int main() {
-    float rep;
-    scanf("%f", &rep);
+    double rep;
+    scanf("%lf", &rep);
     printf("You entered: %.4f", rep);
     return 0;
 }
